[CHORE] 소셜로그인 시 필요한 module 구조 변경 kakao.ts, naver.ts, apple.ts 모두 sns.ts에 지정
